+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### :star2: New features
 
-* Add import of user certificates. After the certificate is imported, it is used automatically when the key is used for authentication.
-* Add support server certificates signed by a trusted authority. CAs are managed with the ca command.
+* Add support for user certificates. After a certificate is imported, it is used automatically when the matching key is used for authentication.
+* Add support for server certificates signed by a trusted authority. CAs are managed with the ca command.
 
 ## v0.4.0